test: module 7 ex02 string type added to tests

+++ module_7/ex02/Array.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Array.hpp                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: husrevakbas <husrevakbas@student.42.fr>    +#+  +:+       +#+        */
+/*   By: huakbas <huakbas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/17 23:32:41 by husrevakbas       #+#    #+#             */
-/*   Updated: 2025/08/20 22:39:55 by husrevakbas      ###   ########.fr       */
+/*   Updated: 2025/08/25 16:03:53 by huakbas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,15 +18,15 @@ template <typename T>
 class Array
 {
 	private:
-		T	*_array;
-		unsigned int	_size;
+		T		*_array;
+		size_t	_size;
 	public:
 		Array<T>();
 		~Array<T>();
 		Array<T>(unsigned int n);
 		Array<T>(const Array<T> &a);
 		Array<T>& operator=(const Array<T> &a);
-		T& operator[](int index);
+		T& operator[](size_t index);
 		unsigned int	size() const;
 		class IndexOutOfRangeException : public std::exception
 		{
